Allow filtering admin user capability type
Oftentimes, custom user roles are defined in WordPress projects to better reflect the editorial workflow. Not always such user roles have the capability to `manage_options` (which is a quite powerful capability).

This change allows filtering the capability name used by the IndexNow admin pages so that WordPress developers working with the plugin have more control.